2️⃣ Fix double events in rating view (#3023)
* 2️⃣ Fix double binding for rating view CTA button

This double binding was causing the unsuccessful events to be fired in the successful case, which we do not want.

* 💄 Cleanup and unify branching code

The PerformMainAction and Dismiss methods are essentially doing the same thing. They are also two many different layers of abstraction. This improves both.

* 💄 DRY code

Lots of repetition in event tracking. This cleans that up a little by extracting shared code into two methods.

* 🔥 Remove inconsistent cleanup

This may be necessary when showing the view multiple times during the same app life cycle.
Right now we don't intend to show it more than once at all after the user interacted with it, so there is no reason to have it.

When in the future we want to show this view more than once, and we expect that to be possible without the app being restarted in the meantime,
we should look this over and make sure the cleanup happens during any moments when this view is hidden, not just in this case.

* :warning: Add warning to throwing property

This is a compromise between complicating both the property and its usage and not giving any indication at all that this will throw when used incorrectly.