🌏Convert Interface Builder strings to Resources.resx (#4123)
## What's this?
Some groundwork to make our iOS app i18n ready. All ( I hope I don't miss any ) the strings that are set in xib, storyboard files are not put inside `Resources.resx` and label init goes to `ViewDidLoad`

**Note**

There are files that contain attributed button that I haven't translated yet. I'll do it when we agree on a solution.

1. CalendarPermissionDeniedViewController.xib
2. NoWorkspaceViewController
3. OutdatedAppViewController

### Relationships
Closes #3872

## Why do we want this?
i18n l10n

## How is it done?

### Why not another way?
We could have a string file for each IB file, but maintaing those are difficult. After discussion with @willsb, we think that having a single `Resources.resx` is the way to go.

### Side effects
Some of the unnecessary line breaks in the middle of the sentence are now removed.

## Review hints
Maybe commit by commit

## :squid: Permissions
Not yet.