🐛 Fix IsBillableAvailable updating the UI (#4707)
## What's this?
Workspace feature IsBillableAvailable now correctly propagates to the UI.

### Relationships
Closes #4694

## Why do we want this?
Because this is the core feature.

## How is it done?
By fixing the subject type.
At first ID used to be initialized to 0, which @simonrozsival deemed dangerous. However, now I'm using `null` value to signify that so far there's no value yet. When it comes, it will be propagated to the UI.

### Why not another way?
This is the easiest way I think.

### Side effects
Now it works. Before it didn't.

### Review hints
Ignore the Reactive Extensions part.
This will disappear when the same change is propagated from develop into the project branch and then into this one. Focus only on the view model part.

## :squid: Permissions
🔟 People.