🐛 GTE: fix startup crash (#4700)
## What's this?
The `await` in `Toggl.Foundation.MvvmCross.App` on line `76` never returned - because the `Current` observable never completed. This led to a crash on `iOS` because initialization didn't create a root VC and iOS had nothing to present to the user.

There must be the same problem on Android, just the app might not crash right away, but the execution must have stopped somehow at the same line.

### Relationships
I'm not sure where exactly this was introduced and why we didn't catch the problem earlier. It was most likely #4618, but I didn't notice a crash back then 🤔.

## Why do we want this?
<!-- Describe the reason for making this change -->

## How is it done?
I added `FirstAsync` and now the user is obtained correctly.

## Review hints
Run the app in iOS simulator.

## :squid: Permissions
🦑