:rainbow: Remove references to MvxColor (#4920)
## What's this?
This completely replaces `MvxColor` with `Shared.Color`

### Relationships
Closes #4186
Ref #4187

## Why do we want this?
MVExit. This touches too many files and we want it to be in develop ASAP.

## How is it done?
We are removing the references to MvxColor. Then we add the necessary code to make our Color struct behave like MvvmCross'. Then we delete the Color plugin.

### Why not another way?
No other way tbh.

### Side effects
Functional test to be sure, but there should be none.

## Review hints
Per commit.

## :squid: Permissions
Only myself.